VEZEPyUniQVerse/pyproject.toml
[build-system]
requires = ["setuptools>=68","wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "veze-uniqverse-portal"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "python-multipart>=0.0.9",
  "jinja2>=3.1",
  "prometheus-client>=0.20",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "httpx>=0.27,<0.28"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "black>=24.8",
  "mypy>=1.11",
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "httpx>=0.27,<0.28",
  "wsproto>=1.2"
]

otel = [
  "opentelemetry-api>=1.27",
  "opentelemetry-sdk>=1.27",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-instrumentation-asgi>=0.48b0",
  "opentelemetry-exporter-otlp>=1.27"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["VEZEPyUniQVerse*", "app*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"app" = [
  "ui/templates/*.html",
  "ui/static/css/*.css",
  "ui/static/js/*.js"
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = false
warn_redundant_casts = false
warn_no_return = false
show_error_codes = true
strict_optional = false

